requirements.txt
# Core geospatial libraries
geopandas>=0.13.0
rasterio>=1.3.0
shapely>=2.0.0
fiona>=1.9.0

# Scientific computing
numpy>=1.24.0
scipy>=1.10.0
scikit-learn>=1.3.0

# Satellite data access
sentinelsat>=1.2.0
requests>=2.31.0

# Configuration and utilities
pyyaml>=6.0
pathlib2>=2.3.0

# Optional: Enhanced processing
opencv-python>=4.8.0
matplotlib>=3.7.0
seaborn>=0.12.0